project/requirements-dev.txt
pytest==7.3.2
factory-boy==3.2.1
pytest-factoryboy==2.5.1
pytest-asyncio==0.21.0
pytest-httpx==0.22.0
flake8==6.0.0
isort==5.12.0
black==23.3.0
bandit==1.7.5

-r requirements.txt
